requirements.txt
rich
click
requests
python-dotenv
openai
google-genai